refactor: historyStatus 히스토리 페이지에서 관리
